Move parsing methods from ffi to ir namespace.
